pyproject.toml
[tool.poetry]
name = "pythonweb-photoshare"
version = "0.1.0"
description = "This is a simple photo-sharing app by Python 17 Web Group-01"
authors = ["Spica12 <example@gmail.com>",
           "Serhii Symbirtsev <example@gmail.com",
           "Andrii Kononov <example@gmail.com",
           "Bogdan Turkot <example@gmail.com",
           "Sergiy Dovgopolyk <example@gmail.com",
]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = {extras = ["all"], version = "0.109.2"}
alembic = "1.13.1"
sqlalchemy = "2.0.27"
asyncpg = "0.29.0"
passlib = "1.7.4"
python-jose = "3.3.0"
fastapi-mail = "1.4.1"
fastapi-users = {extras = ["sqlalchemy"], version = "12.1.3"}
cloudinary = "1.38.0"
psycopg-binary = "3.1.18"
qrcode = "^7.4.2"
pydantic = {extras = ["email"], version = "^2.6.1"}
psycopg2-binary = "^2.9.9"
pillow = "^10.2.0"



[tool.poetry.group.test.dependencies]
pytest = "8.0.0"
pytest-asyncio = "^0.23.5"
httpx = "^0.27.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests"]
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
]
